+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -325,6 +325,17 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
     }
 
     /**
+     * Returns an immutable array containing the last elements that satisfy the [predicate].
+     */
+    public inline fun takeLastWhile(predicate: (element: T) -> Boolean): ImmutableArray<T> {
+        var untilIndex = lastIndex
+        while (untilIndex >= 0 && predicate(values[untilIndex])) {
+            untilIndex--
+        }
+        return takeLast(lastIndex - untilIndex)
+    }
+
+    /**
      * Returns an immutable array containing only the elements matching the given [predicate].
      */
     public inline fun filter(predicate: (element: T) -> Boolean): ImmutableArray<T> {