+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -323,6 +323,17 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
     }
 
     /**
+     * Returns an immutable array containing the last elements that satisfy the [predicate].
+     */
+    public inline fun takeLastWhile(predicate: (element: Short) -> Boolean): ImmutableShortArray {
+        var untilIndex = lastIndex
+        while (untilIndex >= 0 && predicate(values[untilIndex])) {
+            untilIndex--
+        }
+        return takeLast(lastIndex - untilIndex)
+    }
+
+    /**
      * Returns an immutable array containing only the elements matching the given [predicate].
      */
     public inline fun filter(predicate: (element: Short) -> Boolean): ImmutableShortArray {